bump drone-go version

+++ drone/repo/repo_update.go
@@ -52,7 +52,7 @@ var repoUpdateCmd = cli.Command{
 			Name:  "config",
 			Usage: "repository configuration path (e.g. .drone.yml)",
 		},
-		cli.IntFlag{
+		cli.Int64Flag{
 			Name:  "build-counter",
 			Usage: "repository starting build number",
 		},
@@ -85,7 +85,7 @@ func repoUpdate(c *cli.Context) error {
 		ignorePulls  = c.Bool("ignore-pull-requests")
 		cancelPulls  = c.Bool("auto-cancel-pull-requests")
 		cancelPush   = c.Bool("auto-cancel-pushes")
-		buildCounter = c.Int("build-counter")
+		buildCounter = c.Int64("build-counter")
 		unsafe       = c.Bool("unsafe")
 	)
